pyproject.toml
[tool.poetry]
name = "commentupdater"
version = "0.1.0"
description = ""
authors = ["Ziyue Tan <201250093@smail.nju.edu.cn>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.350"
python-dotenv = "^1.0.0"
jupyter = "^1.0.0"
jsonlines = "^4.0.0"
rouge = "^1.0.1"
scikit-learn = "^1.4.0"
pyyaml = "^6.0.1"


[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
